feat(#22): 주문 생성 API 구현
- 포인트 사용 금액을 BigDecimal 로 변경
- 주문하는 상품의 쿠폰 사용 가능 여부 검사 추가
- 주문할 상품의 정보 조회 및 재고 확인 로직 구현
- 고유한 주문 생성 및 주문 상품 엔티티 생성 로직 구현
- 장바구니에서 주문한 상품 삭제 기능 추가

+++ service/order/server/src/main/java/com/sparta/order/server/presentation/dto/OrderDto.java
@@ -1,5 +1,6 @@
 package com.sparta.order.server.presentation.dto;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -15,7 +16,7 @@ public class OrderDto {
 
     private String orderType;
     private List<OrderProductInfo> orderProductInfos = new ArrayList<>();
-    private int pointPrice;
+    private BigDecimal pointPrice;
     private Long addressId;
 
   }
